requirements.txt
Django==3.1.7

celery==5.1.2
redis==3.4.1
django-redis==5.0.0

djangorestframework
djangorestframework-simplejwt
Pillow==8.1.2
redis==3.4.1
requests==2.22.0
urllib3==1.25.7
pytz==2020.1
pandas==1.4.1
gunicorn